libsidplayfp: improve performance of zeroram emulation on ARM platform (increased granularity of MMU, fixed alignment and other fixes).

+++ 3rdparty/sidplayfp/sidplayfp/c64/Banks/SystemRAMBank.h
@@ -36,7 +36,7 @@ class SystemRAMBank : public Bank
 {
 private:
     /** C64 RAM area */
-    unsigned ram[0x10000];
+    uint8_t ram[0x10000];
 
 public:
     /// Initialize RAM with powerup pattern
